requirements.txt
langchain==0.2.14
langchain-community==0.2.12
# langchain-groq==0.1.5
langchain-openai==0.1.20
langchain-core==0.2.32
# langchain-cohere==0.1.8
# langchainhub==0.1.14
# langchain-experimental==0.0.64
# langsmith==0.1.130
rank-bm25==0.2.2
chromadb==0.5.5
fuzzywuzzy==0.18.0
elasticsearch==8.15.0
openpyxl==3.1.5
gradio==4.44
streamlit==1.37.0
python-dotenv==1.0.1
selenium==4.25.0
# datasets==3.0.0
# ragas==0.1.20
scikit-learn==1.5.2
# gensim==4.3.3
fastembed==0.3.6
qdrant-client==1.12.0
icecream==2.1.3
uvicorn==0.30.6
fastapi==0.115.0
psycopg2-binary==2.9.9
# sudo apt-get update
# sudo apt-get install libpq-dev python3-dev
# psycopg2==2.9.9
pandas
openpyxl # Cần thiết để pd.read_excel hoạt động
markdown
beautifulsoup4
seaborn
schedule